fixup! Add auth module
